crf/requirements.txt
pyahocorasick
mecab-python3
python-dotenv
python-crfsuite
